+++ Assets/3.Script/ETC/LivingEntity.cs
@@ -39,7 +39,7 @@ public class LivingEntity : MonoBehaviour, IDamage
         currentState.EnterState(this);
     }
 
-    public virtual void OnDamage(float damage, Vector3 on, Vector3 hitNomal)
+    public virtual void OnDamage(int damage, Vector3 on, Vector3 hitNomal)
     {
         hp -= damage - (damage * Mathf.RoundToInt(def / (def + 50) * 100) * 0.01f);
     }
